Aes.Bits: use a more precise type definition for BitNumber
AFAIK it is not possible with plain Haskell only to constrain a data
type to a pre-defined range of values only.  So if the constructor
function is not given a valid value, we abort immediately.

Also update tests to provide a generic test function for Aes.Bits.bit,
and to test only the range of values of interest.